+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### ⚙️ Engineering
 
-Use `uv` for packaging.
+* Use `uv` for packaging.
 
 
 ## 0.8.3 - September 7, 2024