+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Add reporting of roll-off error state and allow resetting on error.
 
+### 🔧 Fixed
+
+* Restore GS3 status registers and fix addesses for roll-off lockout and error.
+
 
 ## 0.8.4 - November 3, 2024