+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#29](https://github.com/sdss/lvmecp/pull/29) Add a new engineering mode that can be used to bypass the heartbeat and to allow the dome to open during daytime. Part of RORR RID-025.
 * [#30](https://github.com/sdss/lvmecp/pull/30) Prevent the dome from opening during daytime. Close if daytime is detected. Part of RORR RID-025.
+* [#31](https://github.com/sdss/lvmecp/pull/31) Prevent the dome from opening multiple times in a short period of time. RORR RID-019.
 
 ### ✨ Improved