+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next version
+## 1.0.0 - December 25, 2024
 
 ### 🚀 New